server/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg[binary]
python-dotenv
alembic
pydantic[email]
pwdlib[argon2]
redis[async]